fix(core): validate persisted verify command status
